fix issues

+++ backend/app/controllers/categoryController.js
@@ -35,14 +35,21 @@ exports.createCategory = async (req, res) => {
   const { name, description } = req.body;
 
   try {
-    if (!name || !description) {
-      return res.status(400).json({ message: "All fields are required" });
+    if (
+      req.user.user_type !== "super admin" ||
+      req.user.user_type !== "admin"
+    ) {
+      return res.status(401).json({ message: "Unauthorized" });
+    } else {
+      if (!name || !description) {
+        return res.status(400).json({ message: "All fields are required" });
+      }
+
+      const data = { name, description };
+
+      await createCategory(data);
+      return res.status(201).json({ message: "Category created successfully" });
     }
-
-    const data = { name, description };
-
-    await createCategory(data);
-    return res.status(201).json({ message: "Category created successfully" });
   } catch (error) {
     return res.status(500).json({ message: "Internal Server Error" });
   }
@@ -53,14 +60,21 @@ exports.updateCategory = async (req, res) => {
     const { categoryId } = req.params;
     const { name, description } = req.body;
 
-    if (!name || !description) {
-      return res.status(400).json({ message: "All fields are required" });
+    if (
+      req.user.user_type !== "super admin" ||
+      req.user.user_type !== "admin"
+    ) {
+      return res.status(401).json({ message: "Unauthorized" });
+    } else {
+      if (!name || !description) {
+        return res.status(400).json({ message: "All fields are required" });
+      }
+
+      const data = { name, description };
+
+      await updateCategory(data, categoryId);
+      return res.status(200).json({ message: "Category updated successfully" });
     }
-
-    const data = { name, description };
-
-    await updateCategory(data, categoryId);
-    return res.status(200).json({ message: "Category updated successfully" });
   } catch (error) {
     console.log(error);
     return res.status(500).json({ message: "Internal Server Error" });
@@ -71,8 +85,15 @@ exports.deleteCategory = async (req, res) => {
   const { categoryId } = req.params;
 
   try {
-    await deleteCategory(categoryId);
-    return res.status(200).json({ message: "Category deleted successfully" });
+    if (
+      req.user.user_type !== "super admin" ||
+      req.user.user_type !== "admin"
+    ) {
+      return res.status(401).json({ message: "Unauthorized" });
+    } else {
+      await deleteCategory(categoryId);
+      return res.status(200).json({ message: "Category deleted successfully" });
+    }
   } catch (error) {
     return res.status(500).json({ message: "Internal Server Error" });
   }
